feat(report-management): implement report listing functionality with pagination and filtering

+++ src/modules/admin/daily-quest/dto/request/create-daily-quest.request.dto.ts
@@ -4,7 +4,7 @@ import { dailyQuestActionValues } from "../../constants/daily-quest-action-optio
 export const createDailyQuestRequestSchema = z.object({
   code: z.string().trim().min(1).max(191),
   description: z.string().trim().min(1).max(255),
-  karmaReward: z.coerce.number().int().positive(),
+  karmaReward: z.coerce.number().positive(),
   requirement: z.coerce.number().int().positive(),
   action: z.enum(dailyQuestActionValues),
 });
